Update todolist2
I am removing the todolist.my_years() aspect. What Eugenio will do is to grab the set of years from todolist.my_years(). He will then convert it to a list and get a count of the transactions by year. Then he will print a message saying how many transactions for each part of the year. We implemented this database so that it uses datetime, to make the table use less fields. Although we could have used more table columns, the instructor's instructions suggest that we are only allowed one field and should be using python list comprehensions or more advanced sqlite3 queries to find this information as practice to hone our skills.